Move Re2RegexpSplit from functions/lib/Re2Functions.h to functions/prestosql
