+++ velox/functions/lib/Re2Functions.h
@@ -384,70 +384,6 @@ struct Re2RegexpReplace {
   std::string result_;
 };
 
-template <typename TExec>
-struct Re2RegexpSplit {
-  Re2RegexpSplit() : cache_(0) {}
-
-  VELOX_DEFINE_FUNCTION_TYPES(TExec);
-
-  FOLLY_ALWAYS_INLINE void initialize(
-      const std::vector<TypePtr>& /*inputTypes*/,
-      const core::QueryConfig& config,
-      const arg_type<Varchar>* /*string*/,
-      const arg_type<Varchar>* /*pattern*/) {
-    cache_.setMaxCompiledRegexes(config.exprMaxCompiledRegexes());
-  }
-
-  static constexpr int32_t reuse_strings_from_arg = 0;
-
-  void call(
-      out_type<Array<Varchar>>& out,
-      const arg_type<Varchar>& string,
-      const arg_type<Varchar>& pattern) {
-    auto* re = cache_.findOrCompile(pattern);
-
-    const auto re2String = re2::StringPiece(string.data(), string.size());
-
-    size_t pos = 0;
-    size_t lastEnd = 0;
-    const char* start = string.data();
-
-    re2::StringPiece subMatches[1];
-    while (re->Match(
-        re2String,
-        pos,
-        string.size(),
-        RE2::Anchor::UNANCHORED,
-        subMatches,
-        1)) {
-      const auto fullMatch = subMatches[0];
-      const auto offset = fullMatch.data() - start;
-      const auto size = fullMatch.size();
-
-      out.add_item().setNoCopy(
-          StringView(string.data() + lastEnd, offset - lastEnd));
-
-      lastEnd = offset + size;
-      if (UNLIKELY(size == 0)) {
-        pos = lastEnd + 1;
-      } else {
-        pos = lastEnd;
-      }
-    }
-
-    if (LIKELY(pos <= string.size())) {
-      out.add_item().setNoCopy(
-          StringView(string.data() + pos, string.size() - pos));
-    } else {
-      static const StringView kEmptyString(nullptr, 0);
-      out.add_item().setNoCopy(kEmptyString);
-    }
-  }
-
- private:
-  detail::ReCache cache_;
-};
-
 std::shared_ptr<exec::VectorFunction> makeRegexpReplaceWithLambda(
     const std::string& name,
     const std::vector<exec::VectorFunctionArg>& inputArgs,
