+++ velox/functions/lib/tests/Re2FunctionsTest.cpp
@@ -1536,55 +1536,6 @@ TEST_F(Re2FunctionsTest, limit) {
   ASSERT_NO_THROW(evaluate("regexp_like(c0, c2)", data));
 }
 
-TEST_F(Re2FunctionsTest, split) {
-  auto input = makeRowVector({
-      makeFlatVector<std::string>({
-          "1a 2b 14m",
-          "1a 2b 14",
-          "",
-          "a123b",
-      }),
-  });
-  auto result = evaluate("regexp_split(c0, '\\s*[a-z]+\\s*')", input);
-
-  auto expected = makeArrayVector<std::string>({
-      {"1", "2", "14", ""},
-      {"1", "2", "14"},
-      {""},
-      {"", "123", ""},
-  });
-  assertEqualVectors(expected, result);
-
-  result = evaluate("regexp_split(c0, '\\s*\\d+\\s*')", input);
-  expected = makeArrayVector<std::string>({
-      {"", "a", "b", "m"},
-      {"", "a", "b", ""},
-      {""},
-      {"a", "b"},
-  });
-  assertEqualVectors(expected, result);
-
-  // Test for empty matches
-  result = evaluate("regexp_split(c0, '')", input);
-  expected = makeArrayVector<std::string>({
-      {"", "1", "a", " ", "2", "b", " ", "1", "4", "m", ""},
-      {"", "1", "a", " ", "2", "b", " ", "1", "4", ""},
-      {"", ""},
-      {"", "a", "1", "2", "3", "b", ""},
-  });
-  assertEqualVectors(expected, result);
-
-  // Test for another case of empty matches
-  result = evaluate("regexp_split(c0, '\\s*[a-z]*\\s*')", input);
-  expected = makeArrayVector<std::string>({
-      {"", "1", "", "2", "", "1", "4", "", ""},
-      {"", "1", "", "2", "", "1", "4", ""},
-      {"", ""},
-      {"", "", "1", "2", "3", "", ""},
-  });
-  assertEqualVectors(expected, result);
-}
-
 TEST_F(Re2FunctionsTest, parseSubstrings) {
   auto test = [&](const std::string& input,
                   const std::vector<std::string>& expected) {
